feat(claude): add `tools claude spending` alias delegating to ai-spend (#205)
Extract ai-spend's option/action core into a shared registerSpendCommand
helper so `tools claude spending` and `tools ai-spend` share one
implementation with no copy-paste.
